src/test/java/com/saltedfish/community_management/controller/FireSecurityControllerTest.java
package com.saltedfish.community_management.controller;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.RequestBuilder;
import org.springframework.test.web.servlet.ResultActions;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
import org.springframework.test.web.servlet.result.MockMvcResultHandlers;
import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;

import java.sql.Date;

@RunWith(SpringRunner.class)
@SpringBootTest
public class FireSecurityControllerTest {

    @Autowired
    private FireSecurityController fireSecurityController;

    private MockMvc mockMvc;

    @Before
    public void setup(){
        mockMvc = MockMvcBuilders.standaloneSetup(fireSecurityController).build();
    }

    @Test
    public void addFireSecurity() throws Exception {
        RequestBuilder request = MockMvcRequestBuilders.post("/fireSecurity")
                .contentType(MediaType.APPLICATION_JSON)
                .characterEncoding("UTF-8")
                .accept(MediaType.APPLICATION_JSON_UTF8)
                .param("buildId","2")
                .param("createDate",new Date(Date.valueOf("2020-4-1").getTime()).toString())
                .param("checkContent","检查内容2222")
                .param("level","0")
                .param("checkDate",new Date(Date.valueOf("2020-3-20").getTime()).toString());
        ResultActions result = mockMvc.perform(request);
        MvcResult mvcResult = result.andExpect(MockMvcResultMatchers.status().isOk())
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }

    @Test
    public void updateFireSecurity() throws Exception {
        RequestBuilder request = MockMvcRequestBuilders.put("/fireSecurity")
                .contentType(MediaType.APPLICATION_JSON)
                .characterEncoding("UTF-8")
                .accept(MediaType.APPLICATION_JSON_UTF8)
                .param("id","3")
                .param("buildId","2")
                .param("createDate",new Date(Date.valueOf("2020-4-1").getTime()).toString())
                .param("checkContent","检查内容2222")
                .param("level","0")
                .param("checkDate",new Date(Date.valueOf("2020-3-20").getTime()).toString())
                .param("updateDate",new Date(System.currentTimeMillis()).toString());
        ResultActions result = mockMvc.perform(request);
        MvcResult mvcResult = result.andExpect(MockMvcResultMatchers.status().isOk())
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }

    @Test
    public void deleteFireSecurity() throws Exception {
        RequestBuilder request = MockMvcRequestBuilders.delete("/fireSecurity/4")
                .contentType(MediaType.APPLICATION_JSON)
                .characterEncoding("UTF-8")
                .accept(MediaType.APPLICATION_JSON_UTF8);
        ResultActions result = mockMvc.perform(request);
        MvcResult mvcResult = result.andExpect(MockMvcResultMatchers.status().isOk())
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }

    @Test
    public void findFireSecurityById() throws Exception {
        RequestBuilder request = MockMvcRequestBuilders.get("/fireSecurity/3")
                .contentType(MediaType.APPLICATION_JSON)
                .characterEncoding("UTF-8")
                .accept(MediaType.APPLICATION_JSON_UTF8);
        ResultActions result = mockMvc.perform(request);
        MvcResult mvcResult = result.andExpect(MockMvcResultMatchers.status().isOk())
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }

    @Test
    public void findFireSecurity() throws Exception {
        RequestBuilder request = MockMvcRequestBuilders.get("/fireSecurity")
                .contentType(MediaType.APPLICATION_JSON)
                .characterEncoding("UTF-8")
                .accept(MediaType.APPLICATION_JSON_UTF8)
                .param("level","0");
        ResultActions result = mockMvc.perform(request);
        MvcResult mvcResult = result.andExpect(MockMvcResultMatchers.status().isOk())
                .andDo(MockMvcResultHandlers.print())
                .andReturn();
    }

}
